:recycle: [Refactor] User 도메인으로 이동
