+++ src/main/java/com/avab/avab/service/impl/RecreationServiceImpl.java
@@ -69,13 +69,6 @@ public class RecreationServiceImpl implements RecreationService {
         }
     }
 
-    public Page<Recreation> getFavoriteRecreations(User user, Integer page) {
-        Page<RecreationFavorite> favoritesPage =
-                recreationFavoriteRepository.findByUser(user, PageRequest.of(page, 6));
-
-        return favoritesPage.map(RecreationFavorite::getRecreation);
-    }
-
     @Override
     public Page<Recreation> searchRecreations(
             User user,
